@@ -5,7 +5,15 @@
   "keywords": ["yii2", "extension", "longpoll", "longpolling", "comet", "server"],
   "license": "BSD-3-Clause",
   "scripts": {
-    "test": "phpunit --verbose"
+    "test": [
+      "@putenv XDEBUG_MODE=coverage",
+      "Composer\\Config::disableProcessTimeout",
+      "phpunit --coverage-html cover"
+    ],
+    "lint": [
+      "Composer\\Config::disableProcessTimeout",
+      "phpcs"
+    ]
   },
   "authors": [
     {
@@ -14,14 +22,17 @@
     }
   ],
   "require": {
-    "php": "^7.0",
-    "yiisoft/yii2": "~2.0.14",
-    "npm-asset/jquery-longpoll-client": "0.3.0"
+    "php": "^8.1",
+    "yiisoft/yii2": "~2.0.52",
+    "npm-asset/jquery-longpoll-client": "0.3.0",
+    "ext-ctype": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~6.4.0",
     "symfony/process": "*",
-    "yiisoft/yii2-httpclient": "^2.0"
+    "yiisoft/yii2-httpclient": "^2.0",
+    "phpunit/phpunit": "~10.5.38",
+    "squizlabs/php_codesniffer": "^3.11",
+    "ext-curl": "*"
   },
   "repositories": [
     {
@@ -32,6 +43,11 @@
   "autoload": {
     "psr-4": {
       "izumi\\longpoll\\": "src"
+    }
+  },
+  "config": {
+    "allow-plugins": {
+      "yiisoft/yii2-composer": true
     }
   }
 }
